+++ crates/uk-mod/src/pack.rs
@@ -75,27 +75,27 @@ impl std::fmt::Debug for ModPacker {
 #[serde(default)]
 #[serde_as]
 struct InfoJson {
-    name: String,
-    desc: String,
+    name:     String,
+    desc:     String,
     #[serde(deserialize_with = "serde_with::As::<DefaultOnError>::deserialize")]
-    version: String,
+    version:  String,
     platform: String,
-    options: BnpOptions,
+    options:  BnpOptions,
 }
 
 #[derive(Debug, Deserialize, Default)]
 struct BnpOptions {
     #[serde(default)]
-    multi: Vec<BnpOption>,
+    multi:  Vec<BnpOption>,
     #[serde(default)]
     single: Vec<BnpGroup>,
 }
 
 #[derive(Debug, Deserialize)]
 struct BnpOption {
-    name: String,
-    desc: String,
-    folder: PathBuf,
+    name:    String,
+    desc:    String,
+    folder:  PathBuf,
     default: Option<bool>,
 }
 
@@ -141,10 +141,10 @@ impl RequireValue {
 
 #[derive(Debug, Deserialize)]
 struct BnpGroup {
-    name: String,
-    desc: String,
+    name:     String,
+    desc:     String,
     required: Option<RequireValue>,
-    options: Vec<BnpOption>,
+    options:  Vec<BnpOption>,
 }
 
 impl From<BnpGroup> for ExclusiveOptionGroup {
@@ -582,7 +582,7 @@ impl ModPacker {
                     })
                     .transpose()?
                     .unwrap_or_default(),
-                aoc_files: aoc_dir
+                aoc_files:     aoc_dir
                     .map(|aoc| {
                         log::info!("Collecting DLC resources");
                         self_.collect_resources(aoc)